Add postScore module
